+++ repo/lib/my/my_nbr.c
@@ -10,7 +10,7 @@
 int my_getnbr(char const *str)
 {
     int nb = 0;
-    int sign = 1;
+    short sign = 1;
 
     str += *str == '-' && (sign = -1);
     for (int i = 0; ISNUM(str[i]); i++)
